backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pydantic==2.9.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=1.2.0
nsetools>=2.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
